authorization for delete

+++ part4/bloglist/controllers/blogs.js
@@ -37,8 +37,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  // check if token has creators id
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const userId = decodedToken.id
+
+  if (blog.user.toString() === userId.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(200).end()
+  }
+
+  response.status(401).json({
+    error: 'permission denied to delete this blog'
+  })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
